+++ CHANGELOG.md
@@ -31,6 +31,7 @@
   [@chumakoff][]
 * Fixed `if:` scope option when a lambda is passed [#5501][] by [@deivid-rodriguez][]
 * Comment validation adding redundant errors when resource is missing [#5516][] by [@deivid-rodriguez][]
+* Fixed resource filtering by association when the resource has custom primary key [#5446][] by [@wasifhossain][]
 * Fixed "create anoter" checkbox styling. [#5324][] by [@faucct][]
 
 ## 1.3.1 [☰](https://github.com/activeadmin/activeadmin/compare/v1.3.0...v1.3.1)
@@ -387,5 +388,6 @@ Please check [0-6-stable][] for previous changes.
 [@timoschilling]: https://github.com/timoschilling
 [@TimPetricola]: https://github.com/TimPetricola
 [@varyonic]: https://github.com/varyonic
+[@wasifhossain]: https://github.com/wasifhossain
 [@wspurgin]: https://github.com/wspurgin
 [@zorab47]: https://github.com/zorab47